+++ integrations/pgvector/src/haystack_integrations/document_stores/pgvector/document_store.py
@@ -276,7 +276,9 @@ class PgvectorDocumentStore:
             result = cursor.execute(sql_query, params)
         except Error as e:
             self._connection.rollback()
-            detailed_error_msg = f"{error_msg}.\nYou can find the SQL query and the parameters in the debug logs."
+            detailed_error_msg = (
+                f"{error_msg}. Error: {e!r}. \nYou can find the SQL query and the parameters in the debug logs."
+            )
             raise DocumentStoreError(detailed_error_msg) from e
 
         return result
@@ -314,7 +316,9 @@ class PgvectorDocumentStore:
             result = await cursor.execute(sql_query, params)
         except Error as e:
             await self._async_connection.rollback()
-            detailed_error_msg = f"{error_msg}.\nYou can find the SQL query and the parameters in the debug logs."
+            detailed_error_msg = (
+                f"{error_msg}. Error: {e!r}. \nYou can find the SQL query and the parameters in the debug logs."
+            )
             raise DocumentStoreError(detailed_error_msg) from e
 
         return result
@@ -795,7 +799,7 @@ class PgvectorDocumentStore:
         except Error as e:
             self._connection.rollback()
             error_msg = (
-                "Could not write documents to PgvectorDocumentStore. \n"
+                f"Could not write documents to PgvectorDocumentStore. Error: {e!r}. \n"
                 "You can find the SQL query and the parameters in the debug logs."
             )
             raise DocumentStoreError(error_msg) from e
@@ -852,7 +856,7 @@ class PgvectorDocumentStore:
         except Error as e:
             await self._async_connection.rollback()
             error_msg = (
-                "Could not write documents to PgvectorDocumentStore. \n"
+                f"Could not write documents to PgvectorDocumentStore. Error: {e!r}. \n"
                 "You can find the SQL query and the parameters in the debug logs."
             )
             raise DocumentStoreError(error_msg) from e